Moved service store under pkg/api/store/service
