+++ pkg/api/server/userstored/setup.go
@@ -6,12 +6,12 @@ import (
 	"github.com/rancher/norman/store/subtype"
 	"github.com/rancher/norman/types"
 	"github.com/rancher/rancher/pkg/api/customization/workload"
-	"github.com/rancher/rancher/pkg/api/service"
 	"github.com/rancher/rancher/pkg/api/store/ingress"
 	"github.com/rancher/rancher/pkg/api/store/namespace"
 	"github.com/rancher/rancher/pkg/api/store/pod"
 	"github.com/rancher/rancher/pkg/api/store/projectsetter"
 	"github.com/rancher/rancher/pkg/api/store/secret"
+	"github.com/rancher/rancher/pkg/api/store/service"
 	"github.com/rancher/types/apis/project.cattle.io/v3/schema"
 	clusterClient "github.com/rancher/types/client/cluster/v3"
 	"github.com/rancher/types/client/project/v3"
